+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.2
+- Fixed variables without parent objects (non-fields + globals) not generating headers.
+- Added optional style css parameter to the autodoc. 
+  - Prism is the default but custom css can now be provided.
+
 ## 1.0.1
 - Added floating return button in the output autodoc to return readers to the top of the index.
 - Added optional version subtext to the autodoc.